Adds functionality to Primo Barcode lookup addon to enable holds to be placed on items
Why these changes are being introduced:
When we migrated to Alma our Illiad addon for placing holds in Aleph
for LSA scan and deliver service was no longer functional. These updates
to this addon reenable that functionality.

How this addresses that need:
* Adds functionality to place items on hold on behalf of a user
 using the alma API
* user and pickup location are configurable.

Side effects of this change:
None, but this new addon will need to be installed on illiad clients
for staff that require this functionality.

Relevant ticket(s):
https://mitlibraries.atlassian.net/browse/IMP-1767